unit test updates
